+++ R/tm_t_events.R
@@ -516,7 +516,7 @@ tm_t_events <- function(label,
   args <- as.list(environment())
 
   data_extract_list <- list(
-    arm_var = cs_to_des_select(arm_var, dataname = parentname, multiple = TRUE),
+    arm_var = cs_to_des_select(arm_var, dataname = parentname, multiple = TRUE, ordered = TRUE),
     hlt = cs_to_des_select(hlt, dataname = dataname),
     llt = cs_to_des_select(llt, dataname = dataname)
   )
@@ -654,7 +654,7 @@ srv_t_events_byterm <- function(id,
       anl_filtered <- datasets$get_data(dataname, filtered = TRUE)
 
       anl_m <- anl_merged()
-      input_arm_var <- anl_selectors()$arm_var()$select_ordered
+      input_arm_var <- as.vector(anl_m$columns_source$arm_var)
       input_level_term <- c(
         as.vector(anl_m$columns_source$hlt),
         as.vector(anl_m$columns_source$llt)
@@ -721,7 +721,7 @@ srv_t_events_byterm <- function(id,
       my_calls <- template_events(
         dataname = "ANL",
         parentname = "ANL_ADSL",
-        arm_var = anl_selectors()$arm_var()$select_ordered,
+        arm_var = as.vector(anl_m$columns_source$arm_var),
         hlt = if (length(input_hlt) != 0) input_hlt else NULL,
         llt = if (length(input_llt) != 0) input_llt else NULL,
         label_hlt = label_hlt,